requirements.txt
Flask==2.0.2
requests==2.26.0
python-dotenv==0.19.1
feedparser==6.0.8
numpy==1.21.2
python-dateutil==2.8.2